Fix for JPA Query business object, changed queries to typed queries, provided email reminder job with context, and added it to startup service
