+++ timeskip-ejb/src/main/java/be/ehb/scheduler/EmailReminderJobContext.java
@@ -1,0 +1,81 @@
+package be.ehb.scheduler;
+
+import be.ehb.entities.users.UsersWorkLoadActivityBO;
+import be.ehb.mail.BaseMailBean;
+import be.ehb.mail.IMailProvider;
+import be.ehb.storage.IStorageService;
+import org.quartz.JobExecutionException;
+
+import javax.mail.internet.MimeMessage;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+
+/**
+ * Created by Patrick Van den Bussche on 7/05/2017.
+ */
+
+public class EmailReminderJobContext {
+
+    private static final String NEW_LINE = "\n";
+
+    private IMailProvider mp;
+    private IStorageService iss;
+
+    private static void sendMail(StringBuilder content, BaseMailBean bmb, IMailProvider mp) {
+        content.append("Best regards,");
+        content.append(NEW_LINE);
+        content.append("Hr Manager of the Company");
+        content.append(NEW_LINE);
+        content.append("X-X");
+        content.append(NEW_LINE);
+        content.append("*** This is an automatically generated email, please do not reply ***");
+        bmb.setContent(content.toString());
+        MimeMessage mm = mp.composeMessage(bmb);
+        mp.sendMail(mm);
+    }
+
+    void setMp(IMailProvider mp) {
+        this.mp = mp;
+    }
+
+    void setIss(IStorageService iss) {
+        this.iss = iss;
+    }
+
+    void execute() throws JobExecutionException {
+
+        String prevId = null;
+        BaseMailBean bmb = new BaseMailBean();
+
+        List<UsersWorkLoadActivityBO> usersWorkLoadActivityBOList = iss.listUsersWorkloadActivity(new Date());
+
+        if (usersWorkLoadActivityBOList.size() > 0) {
+            StringBuilder content = new StringBuilder();
+            for (UsersWorkLoadActivityBO usersWorkLoadActivityBO : usersWorkLoadActivityBOList) {
+                if (usersWorkLoadActivityBO.getId().equals(prevId)) {
+                    if (prevId != null) {
+                        sendMail(content, bmb, mp);
+                    }
+                    bmb.setTo(usersWorkLoadActivityBO.getEmail());
+                    bmb.setSubject("TimeSkip - Monthly uncompleted tasks");
+                    content.append("Dear ");
+                    content.append(usersWorkLoadActivityBO.getFirstName());
+                    content.append(" ");
+                    content.append(usersWorkLoadActivityBO.getLastName());
+                    content.append(NEW_LINE);
+                    content.append("Please check following uncompleted tasks for last month");
+                    content.append(NEW_LINE);
+                } else {
+                    content.append(new SimpleDateFormat("dd/MM/yyyy").format(usersWorkLoadActivityBO.getDay()));
+                    content.append(" ");
+                    content.append(usersWorkLoadActivityBO.getDescription());
+                    content.append(" ");
+                    content.append(new Long(String.format(usersWorkLoadActivityBO.getLoggedMinutes().toString(), "###")));
+                    content.append(NEW_LINE);
+                }
+            }
+            sendMail(content, bmb, mp);
+        }
+    }
+}
