+++ timeskip-ejb/src/main/java/be/ehb/scheduler/ScheduleService.java
@@ -2,11 +2,14 @@ package be.ehb.scheduler;
 
 
 import be.ehb.factories.ExceptionFactory;
+import be.ehb.mail.IMailProvider;
+import be.ehb.storage.IStorageService;
 import org.quartz.*;
 import org.quartz.impl.StdSchedulerFactory;
 
 import javax.ejb.Singleton;
 import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 
 
 /**
@@ -18,6 +21,11 @@ public class ScheduleService {
 
     private org.quartz.Scheduler sf;
 
+    @Inject
+    private IMailProvider mp;
+    @Inject
+    private IStorageService iss;
+
     public void ScheduleStart() {
         try {
             sf = new StdSchedulerFactory().getScheduler();
@@ -25,27 +33,27 @@ public class ScheduleService {
         catch (SchedulerException ex) {
             throw ExceptionFactory.schedulerNotFoundException("???");
         }
+
+        EmailReminderJobContext mailJob = new EmailReminderJobContext();
+        mailJob.setIss(iss);
+        mailJob.setMp(mp);
+
         JobDetail job = JobBuilder.newJob(EmailReminderJob.class).withIdentity("job1", "group1").build();
         //TODO get date & time from config file, now every first day of month at 10:00 hrs
         //CronTrigger trigger = TriggerBuilder.newTrigger().forJob(job).withSchedule(CronScheduleBuilder.cronSchedule("0 0 10 1 * ? *")).build();
         //TODO for test reason changed to every 3 minutes
         CronTrigger trigger = TriggerBuilder.newTrigger().forJob(job).withSchedule(CronScheduleBuilder.cronSchedule("0 0/3 * * * ? *")).build();
         try {
-            sf.addJob(job, true);
-        }
-        catch (SchedulerException ex) {
-            throw ExceptionFactory.schedulerUnableToAddJobException("???");
-        }
-        try {
             sf.start();
-        }
-        catch (SchedulerException ex) {
+        } catch (SchedulerException ex) {
             throw ExceptionFactory.schedulerUnableToStartException("???");
         }
         try {
+            sf.getContext().put(EmailReminderJob.EMAIL_REMINDER_CONTEXT, mailJob);
             sf.scheduleJob(job, trigger);
         }
         catch (SchedulerException ex) {
+            ex.printStackTrace();
             throw ExceptionFactory.schedulerUnableToScheduleException("???");
         }
     }
